Forking r0.0 for feature/docker-ci

+++ tools/ci/install.sh
@@ -2,26 +2,22 @@
 
 export_stage install && announce_stage
 
-
-
-# Call for dev setup
+# Call for dev setup, see +U_s
 $script_util/parts/init.sh all
 
+# Give private user-script repoo its place
+# TODO: test user-scripts instead/also +U_s +script_mpe
 test -d $HOME/bin/.git || {
+  test "$USER" = "travis" || return 100
+
   rm -rf $HOME/bin || true
   ln -s $HOME/build/bvberkum/script-mpe $HOME/bin
 }
 
 
-sudo ln -s $HOME /srv/home-local
-
-
-# TODO: see +Us Call for dev setup
-$script_util/parts/init.sh all || true
-
-announce "Sourcing env (II)..."
-unset SCRIPTPATH
-. "${BASH_ENV:="$PWD/tools/ci/env.sh"}"
+ci_announce "Sourcing env (II)..."
+unset SCRIPTPATH ci_env_ sh_env_ sh_util_ ci_util_
+. "${BASH_ENV:="$CWD/tools/ci/env.sh"}"
 
 
 #set +o nounset
